Fix a thinko
When creating tables from arrays that are populated iteratively, we need
to add the offset of the current array to the index otherwise we would
be overwriting the lower items in the array. We didn’t hit this before
because most tables we are creating are small and gets loaded in one go,
but not for the palette colors in Segoe UI Emoji.

See https://github.com/khaledhosny/harftex/issues/11